requirements.txt
streamlit>=1.40.0
duckdb>=1.1.0
pandas>=2.2.0
plotly>=5.24.0
openai>=1.50.0
python-dotenv>=1.0.0
